# docs/backend/api/openapi.v1.yaml
openapi: 3.1.0
info:
  title: O(NlogN) API v1
  version: 1.0.0-draft
  description: |
    O(NlogN) API 계약 스켈레톤.
    이 기준선은 루트 메타데이터, 공통 보안, 재사용 가능한 컴포넌트를 정의한다.
  contact:
    name: O(NlogN) API Team
    url: https://api.onlogn.com
servers:
  - url: https://api.onlogn.com
    description: 운영 환경
  - url: http://localhost:3000
    description: 로컬 개발
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/tasks:
    get:
      tags:
        - tasks
      summary: owner task 목록 조회
      operationId: listTasks
      description: |
        인증 사용자를 위한 owner 범위 task 목록.
        기준 정렬은 `created_at desc`, 타이브레이커는 `id desc`다.
      parameters:
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/TaskStatusParam'
        - $ref: '#/components/parameters/TaskVisibilityParam'
        - $ref: '#/components/parameters/TaskGroupIdParam'
        - $ref: '#/components/parameters/TaskDueDateFromParam'
        - $ref: '#/components/parameters/TaskDueDateToParam'
        - $ref: '#/components/parameters/TaskSortParam'
      responses:
        '200':
          description: Task 목록.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '400':
          $ref: '#/components/responses/ProblemResponse'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '403':
          $ref: '#/components/responses/ProblemResponse'
        '500':
          $ref: '#/components/responses/ProblemResponse'
    post:
      tags:
        - tasks
      summary: owner task 생성
      operationId: createTask
      description: 인증 owner 범위에서 task를 생성한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: Task 생성 완료.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '403':
          $ref: '#/components/responses/ProblemResponse'
        '422':
          $ref: '#/components/responses/ValidationProblemResponse'
        '500':
          $ref: '#/components/responses/ProblemResponse'
  /api/v1/tasks/import/google:
    post:
      tags:
        - tasks
      summary: Google Tasks JSON 가져오기
      operationId: importGoogleTasks
      description: 단일 JSON 파일 업로드로 Google Tasks를 1레벨 task로 가져온다.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: 단일 JSON 파일. 최대 2GB.
      responses:
        '201':
          description: 가져오기 완료.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskImportResponse'
        '400':
          $ref: '#/components/responses/ProblemResponse'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '500':
          $ref: '#/components/responses/ProblemResponse'
  /api/v1/tasks/{task_id}:
    get:
      tags:
        - tasks
      summary: owner task 조회
      operationId: getTask
      description: owner 범위 단일 task를 조회한다.
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      responses:
        '200':
          description: Task 상세.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '403':
          $ref: '#/components/responses/ProblemResponse'
        '404':
          $ref: '#/components/responses/ProblemResponse'
        '500':
          $ref: '#/components/responses/ProblemResponse'
    patch:
      tags:
        - tasks
      summary: owner task 수정
      operationId: updateTask
      description: owner 범위 task의 변경 가능한 필드를 수정한다.
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateRequest'
      responses:
        '200':
          description: Task 수정 완료.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '403':
          $ref: '#/components/responses/ProblemResponse'
        '404':
          $ref: '#/components/responses/ProblemResponse'
        '422':
          $ref: '#/components/responses/ValidationProblemResponse'
        '500':
          $ref: '#/components/responses/ProblemResponse'
    delete:
      tags:
        - tasks
      summary: owner task 삭제
      operationId: deleteTask
      description: owner 범위 task를 삭제한다.
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      responses:
        '204':
          description: Task 삭제 완료.
          headers:
            X-Request-Id:
              description: 이 성공 삭제 요청의 추적 id.
              schema:
                type: string
                example: req_01JZ4T8V6W5X4Y3Z2A1
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '403':
          $ref: '#/components/responses/ProblemResponse'
        '404':
          $ref: '#/components/responses/ProblemResponse'
        '500':
          $ref: '#/components/responses/ProblemResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        보호 엔드포인트를 위한 access token 전달 방식.
        auth 엔드포인트에서는 토큰을 JSON 응답 body로 전달한다.
  parameters:
    OffsetParam:
      name: offset
      in: query
      description: 목록 페이지네이션용 0 기반 offset.
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    LimitParam:
      name: limit
      in: query
      description: 목록 페이지네이션용 페이지 크기.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    TaskIdParam:
      name: task_id
      in: path
      description: Task 식별자.
      required: true
      schema:
        type: string
        format: uuid
    TaskStatusParam:
      name: status
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/TaskStatus'
    TaskVisibilityParam:
      name: visibility
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/TaskVisibility'
    TaskGroupIdParam:
      name: group_id
      in: query
      required: false
      description: 그룹 미지정 task를 위한 UUID group id 또는 리터럴 `null`.
      schema:
        oneOf:
          - type: string
            format: uuid
          - type: string
            const: 'null'
    TaskDueDateFromParam:
      name: due_date_from
      in: query
      required: false
      schema:
        type: string
        format: date
    TaskDueDateToParam:
      name: due_date_to
      in: query
      required: false
      schema:
        type: string
        format: date
    TaskSortParam:
      name: sort
      in: query
      description: v1에서는 `created_at desc`만 허용한다.
      required: false
      schema:
        type: string
        enum:
          - created_at desc
        default: created_at desc
...
  schemas:
    TaskStatus:
      type: string
      enum:
        - todo
        - in_progress
        - done
    TaskVisibility:
      type: string
      enum:
        - private
        - public
...
    Problem:
      type: object
      description: RFC9457 problem details 객체.
      required:
        - type
        - title
        - status
        - detail
        - instance
      properties:
        type:
          type: string
          format: uri
          description: 안정적인 problem type URI.
          example: https://api.onlogn.com/problems/resource-not-found
        title:
          type: string
          description: problem type의 짧고 사람이 읽을 수 있는 요약.
          example: Resource Not Found
        status:
          type: integer
          description: HTTP 상태 코드.
          minimum: 100
          maximum: 599
          example: 404
        detail:
          type: string
          description: 이 발생 건에 대한 사람이 읽을 수 있는 설명.
          example: Requested resource does not exist.
        instance:
          type: string
          description: 해당 problem 발생 건을 가리키는 URI 참조.
          example: /api/v1/tasks/req_01JZ4T3Q8R9A2F4M7N6
        request_id:
          type: string
          description: 지원 및 관측을 위한 추적 식별자.
    ValidationErrorItem:
      type: object
      required:
        - field
        - reason
      properties:
        field:
          type: string
          description: 유효하지 않은 요청 필드 이름.
          example: status
        reason:
          type: string
          description: 검증 실패 사유.
          example: "must be one of: todo, in_progress, done"
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            errors:
              type: array
              items:
                $ref: '#/components/schemas/ValidationErrorItem'
    PaginationMeta:
      type: object
      description: 목록 응답용 공통 페이지네이션 메타데이터.
      required:
        - offset
        - limit
        - total
        - sort
      properties:
        offset:
          type: integer
          minimum: 0
          example: 0
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          minimum: 0
          example: 125
        sort:
          type: string
          description: 결정적인 정렬 표현식.
          example: created_at desc, id desc
    ListMeta:
      type: object
      description: 목록 응답을 위한 공통 `meta` 객체 기준선.
      required:
        - pagination
      properties:
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    DataMetaEnvelope:
      type: object
      description: 공통 성공 envelope 기준선(`data` + optional `meta`).
      required:
        - data
      properties:
        data:
          description: 작업 payload, 구체 스키마는 엔드포인트별로 정의된다.
        meta:
          description: 목록/집계 응답용 선택 메타데이터.
          $ref: '#/components/schemas/ListMeta'
...
    Task:
      type: object
      description: Task 리소스 표현.
      required:
        - id
        - owner_user_id
        - group_id
        - title
        - status
        - visibility
        - created_at
        - updated_at
        - due_date
        - start_time
        - end_time
        - tags
        - reference_links
      properties:
        id:
          type: string
          format: uuid
        owner_user_id:
          type: string
          format: uuid
        group_id:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
          default: null
        title:
          type: string
          minLength: 1
        status:
          $ref: '#/components/schemas/TaskStatus'
        visibility:
          $ref: '#/components/schemas/TaskVisibility'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        due_date:
          oneOf:
            - type: string
              format: date
            - type: 'null'
          default: null
        start_time:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
          default: null
        end_time:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
          default: null
        tags:
          type: array
          items:
            type: string
          default: []
        reference_links:
          type: array
          items:
            type: string
            format: uri
          default: []
    TaskCreateRequest:
      type: object
      required:
        - title
      properties:
        group_id:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
        title:
          type: string
          minLength: 1
        status:
          $ref: '#/components/schemas/TaskStatus'
        visibility:
          $ref: '#/components/schemas/TaskVisibility'
        due_date:
          oneOf:
            - type: string
              format: date
            - type: 'null'
        start_time:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        end_time:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        tags:
          type: array
          items:
            type: string
        reference_links:
          type: array
          items:
            type: string
            format: uri
      additionalProperties: false
    TaskUpdateRequest:
      type: object
      properties:
        group_id:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
        title:
          type: string
          minLength: 1
        status:
          $ref: '#/components/schemas/TaskStatus'
        visibility:
          $ref: '#/components/schemas/TaskVisibility'
        due_date:
          oneOf:
            - type: string
              format: date
            - type: 'null'
        start_time:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        end_time:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        tags:
          type: array
          items:
            type: string
        reference_links:
          type: array
          items:
            type: string
            format: uri
      additionalProperties: false
    TaskResponse:
      allOf:
        - $ref: '#/components/schemas/DataMetaEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Task'
      example:
        data:
          id: 123e4567-e89b-12d3-a456-426614174002
          owner_user_id: 123e4567-e89b-12d3-a456-426614174000
          group_id: 123e4567-e89b-12d3-a456-426614174001
          title: Prepare sprint handoff
          status: in_progress
          visibility: private
          created_at: 2026-02-10T09:00:00Z
          updated_at: 2026-02-10T10:30:00Z
          due_date: 2026-02-12
          start_time: 2026-02-10T09:00:00Z
          end_time: null
          tags:
            - planning
            - backend
          reference_links:
            - https://docs.onlogn.com/specs/sprint-handoff
    TaskListResponse:
      allOf:
        - $ref: '#/components/schemas/DataMetaEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Task'
            meta:
              $ref: '#/components/schemas/ListMeta'
      example:
        data:
          - id: 123e4567-e89b-12d3-a456-426614174002
            owner_user_id: 123e4567-e89b-12d3-a456-426614174000
            group_id: 123e4567-e89b-12d3-a456-426614174001
            title: Prepare sprint handoff
            status: in_progress
            visibility: private
            created_at: 2026-02-10T09:00:00Z
            updated_at: 2026-02-10T10:30:00Z
            due_date: 2026-02-12
            start_time: 2026-02-10T09:00:00Z
            end_time: null
            tags:
              - planning
              - backend
            reference_links:
              - https://docs.onlogn.com/specs/sprint-handoff
        meta:
          pagination:
            offset: 0
            limit: 20
            total: 1
            sort: created_at desc, id desc
    TaskImportResult:
      type: object
      required:
        - created_count
        - skipped_count
        - failed_count
      properties:
        created_count:
          type: integer
          minimum: 0
          example: 12
        skipped_count:
          type: integer
          minimum: 0
          example: 3
        failed_count:
          type: integer
          minimum: 0
          example: 1
    TaskImportResponse:
      allOf:
        - $ref: '#/components/schemas/DataMetaEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TaskImportResult'
      example:
        data:
          created_count: 12
          skipped_count: 3
          failed_count: 1
...
  responses:
    ProblemResponse:
      description: RFC9457 problem details 응답.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            authenticationRequired:
              value:
                type: https://api.onlogn.com/problems/authentication-required
                title: Authentication Required
                status: 401
                detail: Access token is missing or expired.
                instance: /api/v1/users/me/req_01JZ4T3Q8R9A2F4M7N6
            permissionDenied:
              value:
                type: https://api.onlogn.com/problems/permission-denied
                title: Permission Denied
                status: 403
                detail: You cannot access another user's private resource.
                instance: /api/v1/groups/01ARZ3NDEKTSV4RRFFQ69G5FAV8/req_01JZ4T4H1X2B3C4D5E6
            resourceNotFound:
              value:
                type: https://api.onlogn.com/problems/resource-not-found
                title: Resource Not Found
                status: 404
                detail: Requested resource does not exist.
                instance: /api/v1/tasks/01ARZ3NDEKTSV4RRFFQ69G5FAV9/req_01JZ4T59K8L7M6N5P4Q
            internalServerError:
              value:
                type: https://api.onlogn.com/problems/internal-server-error
                title: Internal Server Error
                status: 500
                detail: An unexpected error occurred while processing the request.
                instance: /api/v1/groups/req_01JZ4T8V6W5X4Y3Z2A1
                request_id: req_01JZ4T8V6W5X4Y3Z2A1
...
    ValidationProblemResponse:
      description: RFC9457 validation problem details 응답.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblem'
          example:
            type: https://api.onlogn.com/problems/validation-failed
            title: Validation Failed
            status: 422
            detail: Request body validation failed.
            instance: /api/v1/tasks/req_01JZ4T59K8L7M6N5P4Q
            errors:
              - field: status
                reason: 'must be one of: todo, in_progress, done'
              - field: due_date
                reason: must be a valid date in YYYY-MM-DD format